+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.3.0] - 2021-04-02
+### Added
+- Made Certificate Model Name, Certificate Subject DN column name, Certificate Issuer DN column name configurable
+- Added configuration for minimum level of Assurance
+
+### Changed
+- Provisioner will link the Provisioned Certificate to the corresponed CO Person Role(COU)
+
+
 ## [0.2.0] - 2020-10-14
 ### Added
 - Made default DN for the Certificate Authority configurable